requirements.txt
pandas
numpy
matplotlib
streamlit_pandas_profiling